requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
openai==0.11.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.39
pyTelegramBotAPI==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
